ci(www): Add workflow to update copyright annually (#9)
* ci(www): Replace toml action with sed

The toml action wasn't working as expected. It worked with simpler toml
files, but for some reason it just reformats - which trunk wants to
reformat. So going to try find and replace with sed.

* ci(www): Use $GITHUB_ENV instead of env

env: variables were not working with the `data` syntax. Appending variables 
to $GITHUB_ENV via `run` worksas expected

* ci(www): Add workflow to update copyright annually